[♻️refactor]: api 호출 코드 분리
